WIP add integration tests
