+++ frontend/src/metabase/querying/notebook/components/NotebookDataPicker/NotebookDataPicker.tsx
@@ -1,14 +1,7 @@
-import {
-  type MouseEvent,
-  type Ref,
-  forwardRef,
-  useMemo,
-  useState,
-} from "react";
+import { type MouseEvent, useState } from "react";
 import { useLatest } from "react-use";
 import { t } from "ttag";
 
-import { skipToken, useGetCardQuery, useSearchQuery } from "metabase/api";
 import {
   DataPickerModal,
   getDataPickerValue,
@@ -17,13 +10,10 @@ import { METAKEY } from "metabase/lib/browser";
 import { useDispatch, useSelector, useStore } from "metabase/lib/redux";
 import { checkNotNull } from "metabase/lib/types";
 import * as Urls from "metabase/lib/urls";
-import { PLUGIN_EMBEDDING_SDK } from "metabase/plugins";
-import { DataSourceSelector } from "metabase/query_builder/components/DataSelector";
 import { loadMetadataForTable } from "metabase/questions/actions";
-import { getEmbedOptions, getIsEmbedding } from "metabase/selectors/embed";
+import { getIsEmbedding } from "metabase/selectors/embed";
 import { getMetadata } from "metabase/selectors/metadata";
-import type { IconName } from "metabase/ui";
-import { Flex, Icon, Tooltip, UnstyledButton } from "metabase/ui";
+import { Tooltip } from "metabase/ui";
 import * as Lib from "metabase-lib";
 import type { TableId } from "metabase-types/api";
 
@@ -31,8 +21,9 @@ import {
   type NotebookContextType,
   useNotebookContext,
 } from "../Notebook/context";
-import { NotebookCell } from "../NotebookCell";
 
+import { DataPickerTarget } from "./DataPickerTarget";
+import { EmbeddingDataPicker } from "./EmbeddingDataPicker";
 import { getUrl } from "./utils";
 
 interface NotebookDataPickerProps {
@@ -193,169 +184,6 @@ function ModernDataPicker({
   );
 }
 
-type EmbeddingDataPickerProps = {
-  query: Lib.Query;
-  stageIndex: number;
-  table: Lib.TableMetadata | Lib.CardMetadata | undefined;
-  placeholder: string;
-  canChangeDatabase: boolean;
-  isDisabled: boolean;
-  onChange: (tableId: TableId) => void;
-};
-
-function EmbeddingDataPicker({
-  query,
-  stageIndex,
-  table,
-  placeholder,
-  canChangeDatabase,
-  isDisabled,
-  onChange,
-}: EmbeddingDataPickerProps) {
-  const { data: dataSourceCountData, isLoading: isDataSourceCountLoading } =
-    useSearchQuery({
-      models: ["dataset", "table"],
-      limit: 0,
-    });
-
-  const databaseId = Lib.databaseID(query);
-  const tableInfo =
-    table != null ? Lib.displayInfo(query, stageIndex, table) : undefined;
-  const pickerInfo = table != null ? Lib.pickerInfo(query, table) : undefined;
-  const { data: card } = useGetCardQuery(
-    pickerInfo?.cardId != null ? { id: pickerInfo.cardId } : skipToken,
-  );
-
-  const metadata = useSelector(getMetadata);
-  const databases = useMemo(() => {
-    // We're joining data
-    if (!canChangeDatabase) {
-      return [
-        metadata.database(databaseId),
-        metadata.savedQuestionsDatabase(),
-      ].filter(Boolean);
-    }
-
-    /**
-     * When not joining data, we want to use all databases loaded inside `DataSourceSelector`
-     *
-     * @see https://github.com/metabase/metabase/blob/fb25682fe8dbafe2062e37bce832f62440872ab7/frontend/src/metabase/query_builder/components/DataSelector/DataSelector.jsx#L1163-L1168
-     */
-    return undefined;
-  }, [canChangeDatabase, databaseId, metadata]);
-
-  const entityTypes = useSelector(state => getEmbedOptions(state).entity_types);
-
-  if (isDataSourceCountLoading) {
-    return null;
-  }
-
-  const shouldUseSimpleDataPicker =
-    dataSourceCountData != null && dataSourceCountData.total < 100;
-  if (shouldUseSimpleDataPicker) {
-    return (
-      <PLUGIN_EMBEDDING_SDK.SimpleDataPicker
-        filterByDatabaseId={canChangeDatabase ? null : databaseId}
-        selectedEntity={pickerInfo?.tableId}
-        isInitiallyOpen={!table}
-        triggerElement={
-          <DataPickerTarget
-            /**
-             * We try to blur the line between models and tables for embedding users.
-             * this property will change the way icons are displayed in the data picker trigger,
-             * so we need to remove it. Treating it as a table.
-             */
-            getTableIcon={() => "table"}
-            tableInfo={tableInfo}
-            placeholder={placeholder}
-            isDisabled={isDisabled}
-          />
-        }
-        setSourceTableFn={onChange}
-      />
-    );
-  }
-
-  return (
-    <DataSourceSelector
-      key={pickerInfo?.tableId}
-      isInitiallyOpen={!table}
-      databases={databases}
-      canChangeDatabase={canChangeDatabase}
-      selectedDatabaseId={databaseId}
-      selectedTableId={pickerInfo?.tableId}
-      selectedCollectionId={card?.collection_id}
-      databaseQuery={{ saved: true }}
-      canSelectModel={entityTypes.includes("model")}
-      canSelectTable={entityTypes.includes("table")}
-      canSelectMetric={false}
-      canSelectSavedQuestion={false}
-      triggerElement={
-        <DataPickerTarget
-          tableInfo={tableInfo}
-          placeholder={placeholder}
-          isDisabled={isDisabled}
-        />
-      }
-      setSourceTableFn={onChange}
-    />
-  );
-}
-
-type DataPickerTargetProps = {
-  tableInfo?: Lib.TableDisplayInfo;
-  placeholder: string;
-  isDisabled?: boolean;
-  onClick?: (event: MouseEvent<HTMLButtonElement>) => void;
-  onAuxClick?: (event: MouseEvent<HTMLButtonElement>) => void;
-  getTableIcon?: (tableInfo: Lib.TableDisplayInfo) => IconName;
-};
-
-const DataPickerTarget = forwardRef(function DataPickerTarget(
-  {
-    tableInfo,
-    placeholder,
-    isDisabled,
-    onClick,
-    onAuxClick,
-    getTableIcon = defaultGetTableIcon,
-  }: DataPickerTargetProps,
-  ref: Ref<HTMLButtonElement>,
-) {
-  return (
-    <UnstyledButton
-      ref={ref}
-      c="inherit"
-      fz="inherit"
-      fw="inherit"
-      p={NotebookCell.CONTAINER_PADDING}
-      disabled={isDisabled}
-      onClick={onClick}
-      onAuxClick={onAuxClick}
-    >
-      <Flex align="center" gap="xs">
-        {tableInfo && (
-          <Icon name={getTableIcon(tableInfo)} style={{ flexShrink: 0 }} />
-        )}
-        {tableInfo?.displayName ?? placeholder}
-      </Flex>
-    </UnstyledButton>
-  );
-});
-
-function defaultGetTableIcon(tableInfo: Lib.TableDisplayInfo): IconName {
-  switch (true) {
-    case tableInfo.isQuestion:
-      return "table2";
-    case tableInfo.isModel:
-      return "model";
-    case tableInfo.isMetric:
-      return "metric";
-    default:
-      return "table";
-  }
-}
-
 function getModelFilterList(
   { modelsFilterList }: NotebookContextType,
   hasMetrics: boolean,
@@ -366,3 +194,4 @@ function getModelFilterList(
     return modelsFilterList.filter(model => model !== "metric");
   }
 }
+export { DataPickerTarget };
